codes/trajectory/requirements.txt
opencv-python>=4.5
numpy>=1.21
pyserial>=3.5
